Rename typo in dismiss modals
